+++ scripts/download_uniref30.sh
@@ -37,7 +37,20 @@ SOURCE_URL="https://storage.googleapis.com/alphafold-databases/v2.3/UniRef30_202
 BASENAME=$(basename "${SOURCE_URL}")
 
 mkdir --parents "${ROOT_DIR}"
-aria2c "${SOURCE_URL}" --dir="${ROOT_DIR}"
-tar --extract --verbose --file="${ROOT_DIR}/${BASENAME}" \
-  --directory="${ROOT_DIR}"
-rm "${ROOT_DIR}/${BASENAME}"
+
+# Check if the file already exists
+if [[ -f "${ROOT_DIR}/${BASENAME}" ]]; then
+    echo "File ${ROOT_DIR}/${BASENAME} already exists. Skipping download."
+else
+    aria2c "${SOURCE_URL}" --dir="${ROOT_DIR}" --max-connection-per-server=16 --split=16 --min-split-size=1M
+fi
+
+# Check if the extracted folder exists
+EXTRACTED_DIR="${ROOT_DIR}/UniRef30_2021_03"
+if [[ -d "${EXTRACTED_DIR}" ]]; then
+    echo "Folder ${EXTRACTED_DIR} already exists. Skipping extraction."
+else
+    tar --extract --verbose --file="${ROOT_DIR}/${BASENAME}" \
+      --directory="${ROOT_DIR}"
+    rm "${ROOT_DIR}/${BASENAME}"
+fi
